Goldbach in C#
This is the fully functional C# code that will take the data in the data.txt file and write the results into another text file (my_result.txt). If there is no command line argument then the default values (3, 4, 14, 26, and 100) will be printed to the console. Ran on Visual Studio.